Fixed up AccountPage + Lighthouse
Changed the card's size a bit to be easier to read, as well as the font. And formated it to work with tabs. Did do some lighthouse tests, and there is a "bug" of sorts with vuetify mixing with span, and it complains even though all fields and inputs (we have none here) have incorrect aria roles.